requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.36.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.14.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# Weather Data APIs
cdsapi>=0.6.1
requests>=2.31.0
openmeteo-requests>=1.1.0
openmeteo-sdk>=1.6.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
xarray>=2023.8.0
netcdf4>=1.6.4
scipy>=1.11.0

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
wandb>=0.15.0

# Evaluation Metrics
nltk>=3.8.1
rouge-score>=0.1.2
sacrebleu>=2.3.1

# Configuration & Utils
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.1

# Development Tools
jupyter>=1.0.0
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
pre-commit>=3.3.0